Testing foundations:

* Extracted compiler into a separate library
* Exposed span on diagnostics that indicate where the error occurred
* Support for assignments and variables.

+++ Syron/CodeAnalysis/Evaluator.cs
@@ -83,7 +83,7 @@ namespace Syron.CodeAnalysis
                     case BoundBinaryOperatorKind.NotEquals:
                         return !Equals(left, right);
                     case BoundBinaryOperatorKind.Exponentiation:
-                        return Math.Pow((int)left, (int)right);
+                        return (int)Math.Pow((int)left, (int)right);
                     case BoundBinaryOperatorKind.Assignment:
                         return right;
